Add new test grid file.
Everything seems to work with the exception of the extend classes. I
believe this is because of the way the test-grid.scss file is set up.
When the output happens, it takes all classes and adds extend- at the
beginning, but with the way the responsive sections are displayed,
extend only gets added to the first property. I don't believe this is
how the extend really happens when processesing, but I'm not sure how
to do a legitimate test.